+++ src/main/java/org/domiot/backend/MqttMessageHandler.java
@@ -1,0 +1,15 @@
+package org.domiot.backend;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Bean;
+import org.springframework.integration.annotation.ServiceActivator;
+import org.springframework.messaging.MessageHandler;
+
+@Slf4j
+public class MqttMessageHandler {
+    @Bean
+    @ServiceActivator(inputChannel = "mqttInputChannel")
+    public MessageHandler mqttMessageHandler() {
+        return message -> log.info(message.getPayload().toString());
+    }
+}
